Fix starting alpha and beta values
I had it starting out with the wrong values. With this fixed, I quickly hit a
stack depth limit, so for now I've limited it to depth 50, at which point it
returns a draw if no one has won yet.

One idea for an improvement here is to return a heuristic, maybe
own colour pieces / total pieces.